+++ dataset_test.go
@@ -1,0 +1,49 @@
+package dataset
+
+import (
+	"fmt"
+)
+
+func CompareDataset(a, b *Dataset) error {
+	if a.Title != b.Title {
+		return fmt.Errorf("Title mismatch: %s != %s", a.Title, b.Title)
+	}
+
+	if a.Url != b.Url {
+		return fmt.Errorf("Url mismatch: %s != %s", a.Url, b.Url)
+	}
+	if a.Readme != b.Readme {
+		return fmt.Errorf("Readme mismatch: %s != %s", a.Readme, b.Readme)
+	}
+	if a.Author != b.Author {
+		return fmt.Errorf("Author mismatch: %s != %s", a.Author, b.Author)
+	}
+	if a.Image != b.Image {
+		return fmt.Errorf("Image mismatch: %s != %s", a.Image, b.Image)
+	}
+	if a.Description != b.Description {
+		return fmt.Errorf("Description mismatch: %s != %s", a.Description, b.Description)
+	}
+	if a.Homepage != b.Homepage {
+		return fmt.Errorf("Homepage mismatch: %s != %s", a.Homepage, b.Homepage)
+	}
+	if a.IconImage != b.IconImage {
+		return fmt.Errorf("IconImage mismatch: %s != %s", a.IconImage, b.IconImage)
+	}
+	if a.PosterImage != b.PosterImage {
+		return fmt.Errorf("PosterImage mismatch: %s != %s", a.PosterImage, b.PosterImage)
+	}
+	if err := CompareLicense(a.License, b.License); err != nil {
+		return err
+	}
+	if a.Version != b.Version {
+		return fmt.Errorf("Version mismatch: %s != %s", a.Version, b.Version)
+	}
+	if len(a.Keywords) != len(b.Keywords) {
+		return fmt.Errorf("Keyword length mismatch: %s != %s", len(a.Keywords), len(b.Keywords))
+	}
+	// if a.Contributors != b.Contributors {
+	//  return fmt.Errorf("Contributors mismatch: %s != %s", a.Contributors, b.Contributors)
+	// }
+	return nil
+}
