+++ resource_test.go
@@ -10,11 +10,11 @@ import (
 
 func TestResouceHash(t *testing.T) {
 	cases := []struct {
-		r    *Resource
+		r    *Structure
 		hash string
 		err  error
 	}{
-		{&Resource{Format: CsvDataFormat}, "12204ac9a6b596dd42656e7ea7ee3aadf755d92e769f94cc2c08af51aae80889e21b", nil},
+		{&Structure{Format: CsvDataFormat}, "12204ac9a6b596dd42656e7ea7ee3aadf755d92e769f94cc2c08af51aae80889e21b", nil},
 	}
 
 	for i, c := range cases {
@@ -31,10 +31,10 @@ func TestResouceHash(t *testing.T) {
 	}
 }
 
-func TestResourceUnmarshalJSON(t *testing.T) {
+func TestStructureUnmarshalJSON(t *testing.T) {
 	cases := []struct {
 		FileName string
-		result   *Resource
+		result   *Structure
 		err      error
 	}{
 		{"testdata/definitions/airport-codes.json", AirportCodes, nil},
@@ -48,13 +48,13 @@ func TestResourceUnmarshalJSON(t *testing.T) {
 			t.Errorf("case %d couldn't read file: %s", i, err.Error())
 		}
 
-		ds := &Resource{}
+		ds := &Structure{}
 		if err := json.Unmarshal(data, ds); err != c.err {
 			t.Errorf("case %d error mismatch. expected: '%s', got: '%s'", i, c.err, err)
 			continue
 		}
 
-		if err = CompareResources(ds, c.result); err != nil {
+		if err = CompareStructures(ds, c.result); err != nil {
 			t.Errorf("case %d resource comparison error: %s", i, err)
 			continue
 		}
@@ -62,13 +62,13 @@ func TestResourceUnmarshalJSON(t *testing.T) {
 	}
 }
 
-func TestResourceMarshalJSON(t *testing.T) {
+func TestStructureMarshalJSON(t *testing.T) {
 	cases := []struct {
-		in  *Resource
+		in  *Structure
 		out []byte
 		err error
 	}{
-		{&Resource{Format: CsvDataFormat}, []byte(`{"format":"csv","path":"","query":""}`), nil},
+		{&Structure{Format: CsvDataFormat}, []byte(`{"format":"csv","path":"","query":""}`), nil},
 		{AirportCodes, []byte(`{"format":"csv","formatConfig":{"header_row":true},"path":"","query":"","schema":{"fields":[{"name":"ident","type":"string"},{"name":"type","type":"string"},{"name":"name","type":"string"},{"name":"latitude_deg","type":"float"},{"name":"longitude_deg","type":"float"},{"name":"elevation_ft","type":"integer"},{"name":"continent","type":"string"},{"name":"iso_country","type":"string"},{"name":"iso_region","type":"string"},{"name":"municipality","type":"string"},{"name":"gps_code","type":"string"},{"name":"iata_code","type":"string"},{"name":"local_code","type":"string"}]}}`), nil},
 	}
 
@@ -86,11 +86,11 @@ func TestResourceMarshalJSON(t *testing.T) {
 	}
 }
 
-func CompareResources(a, b *Resource) error {
+func CompareStructures(a, b *Structure) error {
 	if a == nil && b == nil {
 		return nil
 	} else if a == nil && b != nil || a != nil && b == nil {
-		return fmt.Errorf("Resource mismatch: %s != %s", a, b)
+		return fmt.Errorf("Structure mismatch: %s != %s", a, b)
 	}
 
 	return nil
